backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mongoengine==0.27.0
pymongo==4.9.0
PyJWT==2.11.0
bcrypt==4.1.1
scikit-learn==1.3.0
numpy==1.24.0
xgboost==2.0.0
drf-yasg==1.21.7
python-json-logger>=2.0.7
setuptools>=65.0.0
